+++ atlas_rag/retriever/hipporag.py
@@ -1,11 +1,12 @@
 from tqdm import tqdm
 import networkx as nx
 import numpy as np
-from atlas_rag.retrieval.embedding_model import BaseEmbeddingModel
+from atlas_rag.vectorstore.embedding_model import BaseEmbeddingModel
 from atlas_rag.llm_generator.llm_generator import LLMGenerator
 from logging import Logger
 from typing import Optional
-from atlas_rag.retrieval.retriever.base import BasePassageRetriever, InferenceConfig
+from atlas_rag.retriever.base import BasePassageRetriever
+from atlas_rag.retriever.inference_config import InferenceConfig
 
 
 class HippoRAGRetriever(BasePassageRetriever):
