+++ atlas_rag/vectorstore/__init__.py
@@ -1,0 +1,1 @@
+from .create_graph_index import create_embeddings_and_index
